use events for torpedo collisions

+++ src/components.rs
@@ -30,7 +30,7 @@ pub struct Armour {
 pub struct Torpedo {
     pub damage: u8,
     pub detonate_radius: f32,
-    pub explosion_radius: u8
+    pub explosion_radius: f32
 }
 
 #[derive(Component)]
